+++ .env.sample
@@ -30,7 +30,7 @@ ETHERSCAN_API_KEY=
 AGGREGATE_QUERY_ID=
 
 # Node for each chain we want to run a sync job on
-NODE_URL_MAINNET=
+NODE_URL_ETHEREUM=
 NODE_URL_GNOSIS=
 NODE_URL_ARBITRUM=
 
@@ -46,3 +46,11 @@ BATCH_DATA_TARGET_TABLE="batch_data"
 
 # Dune api timeout parameter, where we recommend to set it to 600
 DUNE_API_REQUEST_TIMEOUT=600
+
+# The following constants define the upper and lower caps per network
+EPSILON_LOWER_ETHEREUM=
+EPSILON_UPPER_ETHEREUM=
+EPSILON_LOWER_GNOSIS=
+EPSILON_UPPER_GNOSIS=
+EPSILON_LOWER_ARBITRUM=
+EPSILON_UPPER_ARBITRUM=
